Update to calculations controller
